requirements.txt
numpy==1.18.5
Pillow==8.0.1
